product DTO 분리 및 image 로직 추가

+++ ourdus-spring/src/main/java/ourdus/ourdusspring/controller/OnlineClassController.java
@@ -53,7 +53,9 @@ public class OnlineClassController {
     public ApiResult<String> deleteOC(HttpServletRequest req, @PathVariable("class_id") Long classId){
         //Long authorId = Long.valueOf(String.valueOf(jwtService.get(req.getHeader("jwt-auth-token")).get("UserId")));
         Long authorId = 1L;
-        //TODO 작가가 맞는지 로직확인
+//        if(onlineClassService.checkAuthor(authorId, classId))
+//            new ForbiddenException("해당 클래스의 작가가 아니므로 접근할 수 없습니다.");
+        onlineClassService.checkAuthor(authorId, classId);
         onlineClassService.delete(classId);
         return OK("삭제완료");
     }
@@ -62,7 +64,8 @@ public class OnlineClassController {
     public ApiResult<OnlineClassDTO> modifyOC(HttpServletRequest req, @PathVariable("class_id") Long classId, @RequestBody OnlineClassDTO onlineClassDTO){
         //Long authorId = Long.valueOf(String.valueOf(jwtService.get(req.getHeader("jwt-auth-token")).get("UserId")));
         Long authorId = 1L;
-        //TODO 작가가 맞는지 로직확인 필요
+//        if(onlineClassService.checkAuthor(authorId, classId))
+//            new ForbiddenException("해당 클래스의 작가가 아니므로 접근할 수 없습니다.");
         OnlineClass onlineClass = OnlineClass.createBuilder()
                                         .onlineClassDTO(onlineClassDTO)
                                         .build();
